Use togglwrapper instead of TogglPy to access Toggl API
Although TogglPy looks like a more active project, with togglwrapper it is
easier to mock URL requests, as it uses the requests module which works nicely
with pytest together with requests_mock module..

Unfortunately, togglwrapper does not support the reports API, but it is not
that hard to workaround that.

Adding Reports API to togglwrapper and get that merged and release should be
doable.